+++ dashboard/app/controllers/lti_v1_controller.rb
@@ -7,6 +7,9 @@ require "clients/lti_advantage_client"
 require "cdo/honeybadger"
 
 class LtiV1Controller < ApplicationController
+  before_action -> {redirect_to lti_v1_integrations_path, alert: I18n.t('lti.integration.early_access.closed')},
+                if: -> {Policies::Lti.early_access_closed?}, only: :create_integration
+
   # Don't require an authenticity token because LTI Platforms POST to this
   # controller.
   skip_before_action :verify_authenticity_token
@@ -324,7 +327,7 @@ class LtiV1Controller < ApplicationController
       {platform: key, name: value[:name]}
     end
 
-    render lti_v1_integrations_path
+    render template: Policies::Lti.early_access? ? 'lti/v1/integrations/early_access' : 'lti/v1/integrations'
   end
 
   # POST /lti/v1/upgrade_account